Migrate random musings to `/app` directory
