+++ lib/types.ts
@@ -13,6 +13,7 @@ export type Post = {
 };
 
 export type RandomMusing = {
+  excerpt: string;
   content: string;
   date: string;
   slug: string;
